+++ backend/src/test/java/com/team01/backend/domain/comment/controller/CommentControllerTest.java
@@ -14,7 +14,7 @@ import com.team01.backend.domain.post.entity.Post;
 import com.team01.backend.domain.post.repository.PostRepository;
 import com.team01.backend.domain.user.entity.User;
 import com.team01.backend.domain.user.repository.UserRepository;
-import com.team01.backend.domain.user.security.JwtTokenProvider;
+import com.team01.backend.global.security.JwtTokenProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
